[PT] prayer times UI (incomplete)
