+++ android/app/src/main/kotlin/com/developer110/shia_companion/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.developer110.shia_companion
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}
